Add another speculative change to try to fix linking issues on Linux.
In 859b779 some changes were made which improved the situation, but we
still have an issue with the order libraries are specified for
linking. Since AST depends on clangCodeGen, which depends on
LLVMCoverage, add LLVMCoverage here as an explicit dependency. It seems
like that dependency is missing (from CMake's perspective), which is why
we end up with an arbitrary order for those two libraries.