Don't diagnose failures to call symmetrically-typed binary operators as a failure to convert the individual operand, since the operator is likely conceptually generic in some way and the choice of any specific overload is probably arbitrary.
Since we now fall back to a better-informed diagnostics point, take
advantage of this to generate a specialized diagnostic when trying to
compare values of function type with ===.

Fixes rdar://25666129.

This reverts commit 073f4279422163af4f71d15067c6ecf48b6dc54e,
i.e. it reapplies 35ba809fd0665eba2bbe5d0f1cca2f37d00ae620 with a
test fix to expect an extra note in one place.