[constant-prop] Do not crash on frem.
Even though FRem was added as a builtin, this switch was not updated
accordingly. Update the switch and add a test case.

We really should transition all of the switches in this code to be covered
switches so that when builtins are added, it is obvious to know where things
need to be updated...

rdar://26250330