runtime: simplify macro by removing parameters
Remove a few parameters which are easily derived via the C-preprocessor.  The
global symbol name and the implementation names are derived via the
SWIFT_RT_ENTRY_{REF,IMPL} macros in the implementation.  However, we were
repeating the names through the runtime function list.  This simply changes the
macro definitions to derive the name as necessary.  We can restore additional
parameters if the greater flexibility is ever needed.  NFC.