[smoketest] Make sure that we build and test lldb as apart of the OS X smoke tests.
We already do this on the linux smoke tests. From a smoke test perspective, it
would be good to at least be sure that lldb builds/passes basic tests. If we
need to split up the lldb test suite to reduce compile time at some point in the
future, we can explore that.