Build workaround: only consider shims to be protected in Swift code on Darwin.
GNU ld seems to have problems with PC-relative references to
protected functions that aren't calls.  We'll generate that
pattern in Swift code in release builds because of function
merging.  The workaround is to just not tell the compiler
that these symbols can be addressed PC-relative when we might
be using GNU ld, which I've conservatively approximated as
"not Darwin" while I try to figure out what changed to cause this.