rdar://problem/26424505 Import as Member: QoI errors (#2650)
* Add a test that C functions with malformed import-as-member-property API notes don't crash.

* Update ImportAsMember test not to use unprototyped function.

We now reject this in Clang.

* Fix up some more non-prototyped declarations with swift_names.

This is now diagnosed on the Clang side.

* Fix broken API notes uncovered by strengthened Clang-side validation.

* Add test that we still import Clang decls with rejected SwiftName API notes.

Most cases fall out from swift_name validation on the Clang side dropping invalid API notes, though the validation on the Clang side is conservative and misses some cases. We still have Swift-side work to fall back to the original name here.