SILGen: Fix SIL mandatory inlining crash that could occur with public override of an internal property
When we override a property to add a didSet, Sema also synthesizes
a getter that simply delegates to the superclass property getter.

The synthesized getter is marked @_transparent. This means that if
the superclass getter is also transparent, but less visible than
the override, mandatory inlining will crash with an assertion.

To fix this, make sure SILGen does not emit a direct reference to
a superclass method if the current function is marked [fragile]
but the method is not.

Fixes <rdar://problem/26408353>.