Disable this test when lto is enabled.
Right now when we enable lto in swift, we lto the runtime and the compiler. In
the case where we are making dylibs for the runtime this is great, we get better
performance.

On the other hand, there is a bug now that makes it so that the static
libswiftCore also gets the lto bitcode. We do not want users to take this
potential compile time hit from ltoing that archive.

I have not had time to look into this issue yet, so I am xfailing this test when
LTO is enabled.

rdar://26281388