[PrintAsObjC] Handle imported swift_newtype typedefs.
A typedef with the swift_newtype attribute is imported as a struct
wrapping the underlying type instead of just a typealias. If the
underlying type is a bridged type (like String), the newtype struct
is bridged as well. However, we don't want to use that type when
bridging back to Objective-C, because

(1) Objective-C header generation is done too late to fill out the
    _ObjectiveCBridgeable conformance for the type, so if it wasn't
    type-checked then the conformance won't have the original type
    in it.
(2) There's a perfectly good typedef we should be using anyway.

Just use the type as written in Objective-C (instead of crashing).

Finishes rdar://problem/26372925.