Diagnose enum cases renamed to lowercase
When importing foreign enum cases, we create a new case with a lowercase
name, and set the renamed unavailability bits on the imported name, so
they are still visible to the compiler but unavailable, so we can throw
a diagnostic.

Renamed enum cases can be found by digging through a VarDecl in the AST,
and there were no availability checks there. This area probably needs
deeper refactoring, but add an UnavailableInCurrentSwift check for now.

https://bugs.swift.org/browse/SR-1496
rdar://problem/26245512