A workaround for SR-1571
There are probably more robust ways to solve this, but in general it is
not OK to include synthesized constructors in the lookup member set here.  In
the case of classes, you wind up with a different double-diagnostic about the protocol
override and a bogus availability error.  For structs, conformance errors
(if any) are already handled up the chain like classes.