IRGen: Fix Clang type conversion to map Swift.Bool to _Bool instead of signed char
CoreFoundation on Linux expects this. I'm not sure what changed
recently to expose this problem -- and I *am* sure that this
workaround is not the right way of doing it.

However, the public CI on Linux has been broken for a week now
because of this, so let's get it passing again so we can investigate
what changed in Clang/LLVM here.

Fixes <rdar://problem/26444628>.