Reflection: More robust typeref_decoding_objc test
Go through some pains to ensure that we emit an override for
the default initializer, since otherwise we emit a fatalError()
call which ends up creating some closures.

The exact order of the closures changed between master and
swift-3.0-preview-1-branch. It is better not to test this
part at all.