Sema: New fix-its for RawRepresentable conversions
If the expression type is RawRepresentable with an associated
RawValue type of T, and the contextual type is T, suggest adding
a '.rawValue' accessor call.

Also, suggest inserting the fixit in a few more cases, such as
dictionary keys.

This improves upon a previous patch which added a fix-it for the
other direction:

<https://github.com/apple/swift/commit/55bf215feb7f048af1719da2f677ec386166f545>

Fixes <rdar://problem/26470490>.