Reflection fixes for closures and imported types (3.0) (#2736)
* Reflection: Emit metadata for fixed-layout SIL boxes

We were recovering metadata from generic boxes by reading
the instantiated payload metadata from the box's metadata,
but this approach doesn't work for fixed-size boxes, whose
metadata does not store the payload metadata at all.

Instead, emit a capture descriptor with no metadata sources
and a single capture, using the lowered AST type appearing
in the alloc_box instruction that emitted the box.

Since box metadata is shared by all POD types of the same
size, and all single-retainable pointer payloads, the
AST type might not accurately reflect what is actually in
the box.

However, this type is *layout compatible* with the box
payload, at least enough to know where the retainable
pointers are, because after all IRGen uses this type to
synthesize the destructor.

Fixes <rdar://problem/26314060>.

* Reflection: Don't emit associated type records for conformances without associated types

This saves a bit of space. We don't care about conformances per se
in remote reflection, only what the associated types are.

* Reflection: Record builtin and imported types referenced from captures

* Reflection: Don't emit builtin descriptors for imported classes

Previously we would emit both a builtin descriptor and field
descriptor for imported classes, but we only need the latter.

Untangle some code and fix a crash with imported Objective-C
generics in the process.

Fixes <rdar://problem/26498484>.

* Reflection: Emit field metadata for @objc classes

While we can skip emitting metadata for imported classes (clients
should ask the Objective-C runtime instead) it was not correct to
do so for Swift-defined classes that used Objective-C reference
counting.

* Reflection: Emit descriptors for referenced imported protocols

When we encounter a protocol typeref, we have to know if its @objc,
class-bound, or opaque, so make sure we provide the necessary
information when imported protocols are referenced.

* Reflection: Simplify associated type metadata emission

Instead of hooking into nominal type and extension emission
and walking all conformances of those declarations, let's
just directly hook into the logic for emitting conformances.

This fixes an issue where we would apparently emit duplicate
conformances, as well as unnecessary conformances that are
defined elsewhere.

* Reflection: Fix tests for iPhoneOS

* Reflection: More robust typeref_decoding_objc test

Go through some pains to ensure that we emit an override for
the default initializer, since otherwise we emit a fatalError()
call which ends up creating some closures.

The exact order of the closures changed between master and
swift-3.0-preview-1-branch. It is better not to test this
part at all.