Don't emit @objc Fix-Its or diagnostics where @objc inference would do the right job (#2748)
* [Diagnostic verify] Treat "{{none}}" as "no unexpected Fix-Its".

We were using "{{none}}" to mean "no Fix-Its" at all in the diagnostic
verifier, which is useful but narrow. Tweak it's meaning to mean "no
Fix-Its other than the ones we've explicitly checked for", which lets
us verify the exact set of Fix-Its produced as part of a diagnostic
rather than only matching a subset. I'll be landing some tests that
rely on this shortly.

(cherry picked from commit 70f32902525b9070c16d1b6ebc18a444582e4bd3)

* [QoI] Don't emit @objc Fix-Its for cases where @objc will be inferred from a requirement.

This is a follow-up to the change that allowed one to omit @objc (or
the name in an @objc) when it can be inferred by matching a
requirement. There is no point in suggesting that one add @objc if it
will be inferred anyway, since it's just syntactic noise.

(cherry picked from commit 9b24c8391b45074ca0c3074701a492ac78869099)

* QoI: Improve diagnostics for failed @objc inference.

When we cannot infer an @objc name due to an ambiguity, provide a
specific error with Fix-Its for various courses of action (pick a
specific name, make it @nonobjc). Also, be sure to suppress redundant
diagnostics for Objective-C selector checking when checking protocol
conformances: a match-with-renaming will emit the appropriate '@objc'
when it's needed, so we don't need a follow-up diagnostic here.

Finishes rdar://problem/26518216.

(cherry picked from commit a318acb8896c7179e0e6594fcd8f69fea70f4c9c)