SIL: Fix type lowering for generic local functions with captures
We want the generic signature to appear on the outermost
function, after the captures, otherwise we run into trouble
because we have a generic function type in value position,
which is not allowed.

Brings us a step closer to allowing local generic functions
to capture values, but NFC for now.