vim: differentiate between a type and a typename
A typename is a name for a type being defined.  A type may contain optionality
as well.  This ensures that we do not accidentally highlight an invalid name
such as `InvalidName?` as a type name.

Fix the fact that I had previously created a special group for the array.  An
array is a type, so it should be part of the same group.