[build-preset] Add preset buildbot_incremental,tools=RA,llvm-only
This preset just builds llvm via build-script. Its intention is to enable bots
to validate that we are properly configuring LLVM/Clang so that all LLVM/Clang
tests pass.

We have the unfortunate burden of attempting to match LLVM/Clang's build exactly
with build-script. This is a small step towards validating that by ensuring that
when we configure LLVM/Clang we can at least run their tests.

For expediency, I am just going to trigger the LLVM/Clang tests on the bots
without using build-script-impl since there are a bunch of large changes
occuring there.

rdar://26559922